interpolation search

+++ 0x1E-search_algorithms/102-interpolation.c
@@ -1,13 +1,13 @@
 #include "search_algos.h"
 
 int interpolation_search(int *arr, size_t size, int value) {
-    int low = 0, high = (size - 1);
+    size_t low = 0, high = (size - 1);
     while (low <= high && value >= arr[low] && value <= arr[high]) {
         if (low == high) {
             if (arr[low] == value) return low;
             return -1;
         }
-        int pos = low + (((double)(high - low) / (arr[high] - arr[low])) * (value - arr[low]));
+        size_t pos = low + (((double)(high - low) / (arr[high] - arr[low])) * (value - arr[low]));
         if (arr[pos] == value) return pos;
         if (arr[pos] < value) low = pos + 1;
         else high = pos - 1;
